refactor spring-xsuaa autoconfig
