+++ spring-xsuaa/src/main/java/com/sap/cloud/security/xsuaa/autoconfiguration/XsuaaAutoConfiguration.java
@@ -5,7 +5,13 @@
  */
 package com.sap.cloud.security.xsuaa.autoconfiguration;
 
+import com.sap.cloud.security.config.CredentialType;
+import com.sap.cloud.security.xsuaa.XsuaaServiceConfiguration;
+import com.sap.cloud.security.xsuaa.XsuaaServiceConfigurationDefault;
+import com.sap.cloud.security.xsuaa.XsuaaServicePropertySourceFactory;
 import com.sap.cloud.security.xsuaa.extractor.TokenUtil;
+import com.sap.cloud.security.xsuaa.mtls.ServiceClientException;
+import com.sap.cloud.security.xsuaa.mtls.SpringHttpClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
@@ -13,17 +19,13 @@ import org.springframework.boot.autoconfigure.condition.AllNestedConditions;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Conditional;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.PropertySource;
+import org.springframework.context.annotation.*;
+import org.springframework.core.type.AnnotatedTypeMetadata;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.web.client.RestOperations;
 import org.springframework.web.client.RestTemplate;
 
-import com.sap.cloud.security.xsuaa.XsuaaServiceConfiguration;
-import com.sap.cloud.security.xsuaa.XsuaaServiceConfigurationDefault;
-import com.sap.cloud.security.xsuaa.XsuaaServicePropertySourceFactory;
+import javax.annotation.Nonnull;
 
 /**
  * {@link EnableAutoConfiguration Auto-configuration} for default beans used by
@@ -78,16 +80,41 @@ public class XsuaaAutoConfiguration {
 	}
 
 	/**
-	 * Creates a {@link RestOperations} instance if the application has not yet
-	 * defined any yet.
+	 * Creates a {@link RestOperations} instance if the application has not
+	 * defined any.
 	 *
 	 * @return the {@link RestOperations} instance.
 	 */
 	@Bean
+	@Conditional({OnNotX509CredentialTypeCondition.class})
 	@ConditionalOnMissingBean
 	public RestOperations xsuaaRestOperations() {
-		logger.info("auto-configures RestOperations for xsuaa requests");
+		logger.warn("In productive environment provide a well configured client secret based RestOperations bean");
 		return new RestTemplate();
+	}
+
+	/**
+	 * Creates a certificate based {@link RestOperations} instance if the application has not
+	 * defined any.
+	 *
+	 * @return the {@link RestOperations} instance.
+	 */
+	@Bean
+	@ConditionalOnProperty(prefix = "xsuaa", name = "credential-type", havingValue = "x509")
+	@ConditionalOnMissingBean
+	public RestOperations xsuaaMtlsRestOperations(XsuaaServiceConfiguration xsuaaServiceConfiguration) throws ServiceClientException {
+		logger.warn("In productive environment provide a well configured certificate based RestOperations bean");
+		return SpringHttpClient.getInstance().create(xsuaaServiceConfiguration.getClientIdentity());
+	}
+
+	private static class OnNotX509CredentialTypeCondition implements Condition {
+		@Override
+		public boolean matches(ConditionContext context, @Nonnull AnnotatedTypeMetadata metadata) {
+			CredentialType credentialType = CredentialType
+					.from(context.getEnvironment().getProperty("xsuaa.credential-type"));
+			return credentialType == CredentialType.BINDING_SECRET || credentialType == CredentialType.INSTANCE_SECRET
+					|| credentialType == null;
+		}
 	}
 
 }
